+++ ErsatzTV.Core/Metadata/MediaItemTemplateDataKey.cs
@@ -2,10 +2,11 @@ namespace ErsatzTV.Core.Metadata;
 
 public static class MediaItemTemplateDataKey
 {
-    public static readonly string Resolution = "Resolution";
-
+    public static readonly string Resolution = "MediaItem_Resolution";
     public static readonly string Duration = "MediaItem_Duration";
     public static readonly string StreamSeek = "MediaItem_StreamSeek";
+    public static readonly string Start = "MediaItem_Start";
+    public static readonly string Stop = "MediaItem_Stop";
 
     // common
     public static readonly string Path = "MediaItem_Path";